Move launch_pages_internal to a new file
